% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translated_levels.R
\name{add_translated_levels}
\alias{add_translated_levels}
\title{Add translated levels to a list object}
\usage{
add_translated_levels(list_obj, ...)
}
\arguments{
\item{list_obj}{A list object containing the levels to be translated.}

\item{...}{Additional arguments passed to the \code{translated_levels} function.}
}
\value{
A modified version of a variable description list object
with the translated levels added.
}
\description{
Translate the \code{levels} vector for a categorical factor and add it to a list object.
}
\examples{
p_domain <- list(
 levels = c("Coca", "Contraband", "Education")
)
add_translated_levels(p_domain, variable = "protest_domain", uc_translation)
}
