% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-var-name-table.R
\name{append_to_var_name_table}
\alias{append_to_var_name_table}
\title{Append new entries to variable_name_table}
\usage{
append_to_var_name_table(table, extension, overwrite = FALSE)
}
\arguments{
\item{table}{A variable translation table with one column named \code{language}}

\item{extension}{A matching variable translation table with one
column named \code{language} that has the same languages, not necessarily
in the same order. (The extension may have additional languages, which
will be dropped.)}

\item{overwrite}{A flag indicating whether entries that overlap between
the two tables should be overwritten by the value in extension.
Default value is FALSE, indicating no change should be made.}
}
\value{
A longer variable translation table
}
\description{
Append new entries to variable_name_table
}
\examples{
n_trans_table <- tibble::tribble(
 ~language, ~date, ~age, ~n, ~n_state_perp, ~n_state_victim, ~n_state_separate, ~deaths,
 "en", "Date", "Age", "Confirmed", "State Perp", "State Victim", "Sep from State", "Deaths",
 "es", "Fecha", "Edad", "Confirmado", "Perp x Estado", "Víctima Estatal",
   "Ajeno del Estado", "Muertes",
 "r_variable", "date", "age", "n", "n_state_perp", "n_state_victim",
   "n_state_separate", "deaths")
append_to_var_name_table(uc_var_table, n_trans_table)
append_to_var_name_table(uc_var_table, gss_var_table)[c(1,6,71)]
}
