% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blank_translation_table.R
\name{create_blank_translation_table}
\alias{create_blank_translation_table}
\title{Create one blank translation table}
\usage{
create_blank_translation_table(
  dataframe,
  variable,
  source_lang = tcats$source_lang,
  dest_language_list = tcats$dest_lang_list
)
}
\arguments{
\item{dataframe}{Dataframe whose columns are the variables}

\item{variable}{Single variable among the columns, passed as string}

\item{source_lang}{Code for source language; will head left column of output}

\item{dest_language_list}{Codes, as a list of strings, to be translated to}
}
\value{
Data frame with possible values for the variable in the source
  language and blank to fill in translations
}
\description{
Creates a blank translation table for a single variable in a data frame.
}
\details{
Left column, named by `source_lang` has the original values. If the
variable is a factor, these are the levels of the factor. If it's a
character variable, these are the unique values of `variable` in
`dataframe`.

Each language in the `dest_language_list` gets its own column. However,
if the `source_lang` is in this list, it is removed. Duplicate values
receive a warning.
}
\examples{
create_blank_translation_tables(forcats::gss_cat, "relig",
    "en", c("fr", "cz"))
}
