% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blank_translation_table.R
\name{create_blank_translation_tables}
\alias{create_blank_translation_tables}
\title{Create a composite translation table}
\usage{
create_blank_translation_tables(
  dataframe,
  variable_list,
  source_lang = tcats$source_lang,
  dest_language_list = tcats$dest_lang_list,
  incl_header = TRUE,
  combine_tables = TRUE
)
}
\arguments{
\item{dataframe}{Dataframe whose columns are the variables}

\item{variable_list}{List of variables}

\item{source_lang}{Code for source language; will head left column of output}

\item{dest_language_list}{Codes, as a list of strings, to be translated to}

\item{incl_header}{If true, put the name of the variable at the top of each
translation table.}

\item{combine_tables}{If true, combine all the translation tables into a
single data frame. If false, return a single list of translation tobles,
with the variable names as the names of the list elements.}
}
\value{
A single dataframe or a list of dataframes
}
\description{
Creates a blank translation table for a multiple variables in a data frame.
}
\details{
If \code{combine_tables} is TRUE, all the translation tables are stacked and
bound together in a single table. If FALSE, they are returned in the
preferred \code{transcats} form, a list of translation tables.

The left column of each table, named by \code{source_lang} has the original values. If the
variable is a factor, these are the levels of the factor. If it's a
character variable, these are the unique values of \code{variable} in
\code{dataframe}.

Each language in the \code{dest_language_list} gets its own column. However,
if the \code{source_lang} is in this list, it is removed. Duplicate values
receive a warning.
}
\examples{
create_blank_translation_tables(forcats::gss_cat, c("marital", "race", "relig"),
    "en", c("fr", "cz"))
create_blank_translation_tables(vcd::DanishWelfare, c("Status", "Urban"),
                                source_lang = "en",
                                dest_language_list = c("en-complete", "dk"))
}
