% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{dest_lang_list}
\alias{dest_lang_list}
\title{Report the active destination language list for translation}
\usage{
dest_lang_list()
}
\description{
Report the active destination language list for translation
}
