% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation-functions.R
\name{get_translation_variables}
\alias{get_translation_variables}
\title{Get list of variables in the active translation table}
\usage{
get_translation_variables()
}
\value{
A character vector with the names of the variables in the
active translation table.
}
\description{
Returns a list of variables in the active translation table.
}
\examples{
set_active_translation_table(uc_translation)
get_translation_variables()
}
