% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss_translation}
\alias{gss_translation}
\title{Translation table for gss_cat}
\format{
\subsection{\code{gss_translation}}{

A list of tibbles, each with three columns named "en", "es", "fr".
Their names correspond to the categorical variables involved:
\code{marital}, \code{race}, \code{rincome}, \code{partyid}, \code{relig}, \code{denom}.
}
}
\usage{
gss_translation
}
\description{
A translation table for the variable gss_cat, a data source from the
General Social Survey that is included in the package \code{forcats}
and used for examples in \code{transcats}.
}
\keyword{datasets}
