% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_combined_translation_table.R
\name{parse_combined_translation_table}
\alias{parse_combined_translation_table}
\title{Reformat a combined translation table into a standard one}
\usage{
parse_combined_translation_table(combined_table)
}
\arguments{
\item{combined_table}{A combined translation table with as
many columns as languages involved. This table must be
sectioned off for each variable. Each header line takes the form
"variable_name", "--", "--", with as many "--"s as needed
to cover all but the first column.}
}
\value{
A translation table list, \code{translation_list},
each member of which is a translation table for a
given categorical variable.
}
\description{
Part of the import sequence for a combined translation table,
which can be produced by \code{create_blank_translation_table} and
edited externally (for example, as a CSV file). This function
takes a single data frame and splits it into as many smaller
translation tables as are indicated by the number. of headers.
}
\details{
It returns the translation table list used as the main
translation table format by \code{transcats}.
}
\examples{
gss_translation_combined <- readr::read_csv(
                       fs::path_package("extdata", "gss_cat_transtable_complete.csv",
                                        package = "transcats"))
gss_translation <- parse_combined_translation_table(gss_translation_combined)
}
