% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable-lang-es.R
\name{reactable_lang_en}
\alias{reactable_lang_en}
\title{Provide English-language text for interactive Reactable interface elements}
\usage{
reactable_lang_en(plural_noun = "rows")
}
\arguments{
\item{plural_noun}{A word in English, formulated as a plural noun
that denotes what the rows of the table refer to, such as
"observations", "events", etc.
By default, it is set to "rows."}
}
\value{
A specialized list of interface text for the Reactable
package to use. Follows the format of \code{\link[reactable:reactableLang]{reactable::reactableLang()}}
}
\description{
Provide English-language text for interactive Reactable interface elements
}
\examples{
reactable_lang_en_1 <- reactable_lang_es("events")
}
