% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable-lang-es.R
\name{reactable_lang_es}
\alias{reactable_lang_es}
\title{Provide Spanish-language text for interactive Reactable interface elements}
\usage{
reactable_lang_es(plural_noun = "filas")
}
\arguments{
\item{plural_noun}{A word in Spanish, formulated as a plural noun
that denotes what the rows of the table refer to, such as
"observaciones" (observations), "eventos" (events), etc.
By default, it is set to "filas" (rows).}
}
\value{
A specialized list of interface text for the Reactable
package to use. Follows the format of \code{\link[reactable:reactableLang]{reactable::reactableLang()}}
}
\description{
Provide Spanish-language text for interactive Reactable interface elements
}
\examples{
reactable_lang_es_1 <- reactable_lang_es("eventos")
}
