% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation-functions.R
\name{set_active_translation_table}
\alias{set_active_translation_table}
\title{Change the active translation table}
\usage{
set_active_translation_table(transtable)
}
\arguments{
\item{transtable}{A list of translation tables, named
by the variables involved.}
}
\value{
Former value of the parameter.
}
\description{
Changes the value of tcats$translation_table.
}
