% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{set_dest_lang}
\alias{set_dest_lang}
\title{Change the active destination language for translation}
\usage{
set_dest_lang(langcode = "es")
}
\arguments{
\item{langcode}{Code for the language you are choosing, typically
a two- or three-letter string.}
}
\value{
Former value of the parameter.
}
\description{
Changes the value of tcats$dest_lang.
}
