% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{set_dest_lang_list}
\alias{set_dest_lang_list}
\title{Change the active destination language list for multiple translations}
\usage{
set_dest_lang_list(langlist = c(tcats$dest_lang))
}
\arguments{
\item{langcode}{Code for the language you are choosing, typically
a two- or three-letter string.}
}
\value{
Former value of the parameter.
}
\description{
Changes the value of tcats$dest_lang_list.
}
