% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{set_title_lang}
\alias{set_title_lang}
\title{Change the active language for names of variables}
\usage{
set_title_lang(langcode = "en")
}
\arguments{
\item{langcode}{Code for the language you are choosing, typically
a two- or three-letter string.}
}
\value{
Former value of the parameter.
}
\description{
Changes the value of tcats$title_lang.
}
