% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{set_var_name_table}
\alias{set_var_name_table}
\title{Change the active variable name table}
\usage{
set_var_name_table(vartable)
}
\arguments{
\item{vartable}{A horizontal translation table whose column names
correspond to variables}
}
\value{
Former value of the parameter.
}
\description{
Changes the value of tcats$var_name_table, the currently active
table for translating variable names.
}
\examples{
set_var_name_table(gss_var_table)
}
