% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sr_levels}
\alias{sr_levels}
\title{State Responsibility Levels}
\format{
`sr_levels` is a vector with six values.
}
\source{
<https://ultimateconsequences.github.io/>
}
\usage{
sr_levels
}
\description{
A vector ordering the levels for the variable `state_responsibility` in
the Ultimate Consequences database.
}
\details{
sr_levels <- c("Perpetrator", "Victim", "Involved", "Separate",
               "Unintentional", "Unknown")
}
\keyword{datasets}
