% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcats-package.R
\docType{package}
\name{transcats-package}
\alias{transcats}
\alias{transcats-package}
\title{transcats: Translate Categorical Data Using Translation Tables}
\description{
Translate the variable names and titles and values of dataframes containing containing categorical and other data. These translations are made using editable translation tables to allow users to edit and verify translations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CarwilB/transcats}
  \item Report bugs at \url{https://github.com/CarwilB/transcats/issues}
}

}
\author{
\strong{Maintainer}: Carwil Bjork-James \email{c.bjork-james@vanderbilt.edu} (\href{https://orcid.org/0000-0002-4977-0458}{ORCID}) [copyright holder]

}
\keyword{internal}
