% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translated_levels.R
\name{translated_levels}
\alias{translated_levels}
\title{Translated levels}
\usage{
translated_levels(
  levels_vector,
  variable,
  translation_table,
  dest_lang = "es",
  source_lang = "en"
)
}
\arguments{
\item{levels_vector}{One-dimensional vector of levels of a factor}

\item{variable}{The variable involved, passed as a string.}

\item{translation_table}{Consolidated translation table to be used
(translation_table$variable should be a translation table for
the variable involved).}

\item{dest_lang}{Language to be translated to, typically as a short code.}

\item{source_lang}{Language to be translated from, typically as a short code.}
}
\value{
A one-dimensional levels vector for the factor.
}
\description{
Translate the levels vector for a categorical factor.
}
\examples{
sr_levels_es <- sr_levels \%>\% translated_levels("state_responsibility")
}
