% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_name.R
\name{var_lang_str}
\alias{var_lang_str}
\title{Language-specific names for variables (shortened call)}
\usage{
var_lang_str(...)
}
\arguments{
\item{...}{Name of the variable, passed as a string, in quotes}
}
\value{
A text string that is the title of the variable
}
\description{
This functions provides a shortened call to variable_name using
preset values for the language and variable name table.
}
\examples{
var_lang_str("tvhours")


}
