% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_name.R
\name{var_name_language_available}
\alias{var_name_language_available}
\title{Check if a language is available in a variable name table}
\usage{
var_name_language_available(
  lang = tcats$title_lang,
  name_table = tcats$var_name_table
)
}
\arguments{
\item{lang}{A language code.}

\item{name_table}{A variable name table. Defaults to the active var_name_table.}
}
\value{
Logical true or false
}
\description{
This function returns a simple logical result, TRUE or FALSE, as to
whether a given language appears in the active variable name table, or in
a language table passed as the second parameter.
}
\examples{
var_name_language_available("fr") # examines active variable table
var_name_language_available("es", uc_var_table)
}
