% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_name.R
\name{variable_name}
\alias{variable_name}
\title{Language-specific names for variables}
\usage{
variable_name(
  variable,
  lang = tcats$title_lang,
  name_table = tcats$var_name_table
)
}
\arguments{
\item{variable}{Name of the variable, without quotes}

\item{lang}{Language to be used; use `r_variable` for the variable itself.}

\item{name_table}{A horizontal translation table whose column names correspond to variables}
}
\value{
A text string that is the title of the variable, or the name of the variable in r.
}
\description{
Language-specific names for variables
}
\examples{
variable_name_from_string(relig, "es", gss_var_table)
variable_name_from_string(partyid, "en", gss_var_table)

}
