% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_name.R
\name{variable_name_from_string}
\alias{variable_name_from_string}
\title{Language-specific names for variables}
\usage{
variable_name_from_string(
  variable,
  lang = tcats$title_lang,
  name_table = tcats$var_name_table
)
}
\arguments{
\item{variable}{Name of the variable, passed as a string, in quotes}

\item{lang}{Language to be used; use `r_variable` for the variable itself.}

\item{name_table}{A horizontal translation table whose column names correspond to variables}
}
\value{
A text string that is the title of the variable, or the name of the variable in r.
}
\description{
This function outputs a string naming a variable based on a multilingual
translation table. If you want to output the names of multiple variables at
once, you can use this function with map() (and you cannot use variable_name()).
}
\examples{
set_var_name_table(gss_var_table)
variable_name_from_string("relig", lang="es")
variable_name_from_string("partyid", lang="en")
purrr::map(colnames(forcats::gss_cat), var_lang_str)


}
