% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_name.R
\name{variable_names_vector}
\alias{variable_names_vector}
\title{Produce a list of display names for each variables}
\usage{
variable_names_vector(dataframe)
}
\arguments{
\item{dataframe}{Dataframe whose columns are the variables to be listed.}

\item{lang}{Language to be used; use `r_variable` for the variable itself.}

\item{name_table}{A horizontal translation table whose column names correspond to variables}
}
\value{
A vector of strings wit tehe display names.
}
\description{
Using the variable name table, produce a list of display names for each
variable, in the language specified.
}
\examples{
variable_names_vector(gss_cat)
}
