% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation-available.R
\name{which_translation_available}
\alias{which_translation_available}
\title{Show values with available translations}
\usage{
which_translation_available(
  variable_name,
  translation_table = tcats$translation_table,
  dest_lang = tcats$dest_lang,
  source_lang = tcats$source_lang,
  verbose = FALSE
)
}
\arguments{
\item{variable_name}{The name of the variable to be checked,
as a string.}

\item{translation_table}{A translation table, by default
this is the active translation table.}

\item{dest_lang}{The destination language code, by default
this is the current source language, set by set_dest_lang().}

\item{source_lang}{The source language code, by default
this is the current source language, set by set_source_lang().}

\item{verbose}{If TRUE, prints messages about the process.}
}
\value{
A character vector with the values that have translations
available.
}
\description{
\code{which_translation_available} returns the values for which
there is an available translation in the translation table.

Note that unlike functions like \code{translation_available_for},
\code{which_translation_available} does not check the values
within a dataframe or vector, but looks directly at the
translation table and returns all possible values that
can be translated.
}
