% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss_var_table}
\alias{gss_var_table}
\title{Variable name table for gss_cat}
\format{
\subsection{\code{gss_var_table}}{

A data frame with 4 rows and 10 columns. The rows correspond to
English, Spanish, French and the r_variable. The final row matches the
names of the columns. Rows are indexted by language codes in the variable
language.
}
}
\source{
Created as part of \code{transcats}
}
\usage{
gss_var_table
}
\description{
A variable name translation table for the variable gss_cat,
a data source from the General Social Survey that is included
in the package \code{forcats} and used for examples in \code{transcats}.
}
\keyword{datasets}
