% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation-functions.R
\name{translated_join}
\alias{translated_join}
\title{Translated Join}
\usage{
translated_join(
  dataframe,
  variable,
  translation_table = tcats$translation_table,
  dest_lang = tcats$dest_lang,
  source_lang = tcats$source_lang
)
}
\arguments{
\item{dataframe}{Data table being changed}

\item{variable}{Name of the variable to be translated, as a string.}

\item{translation_table}{Consolidated translation table to be used
(translation_table$variable should be a translation table for
the variable involved).}

\item{dest_lang}{Language to be translated to, typically as a short code.}

\item{source_lang}{Language to be translated from, typically as a short code.}

\item{variables}{A list of variables in c("firstvar", "secondvar") format.}
}
\value{
A modified version of the data table with extra columns
}
\description{
Adds one column (`translated_join`) or several columns (`translated_join_vars`)
to the right side of a data table, filled with translations of one or several
categorical variables.
}
\details{
Translated variables take the name of the original variable, plus an underscore,
and the code of the destination language.
}
\examples{
sr_levels_es <- sr_levels \%>\% translated_levels("state_responsibility")
}
