% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uc_translation}
\alias{uc_translation}
\title{Ultimate Consequences Translation Table}
\format{
\code{uc_translation} is a list of eight tibbles, each with two columns named "en" and "es".
Their names correspond to the categorical variables involved:
\code{protest_domain}, \code{pres_admin}, \code{dec_affiliation}, \code{perp_affiliation},
\code{cause_death}, \code{state_responsibility}, \code{intentionality}, \code{event_title}
}
\usage{
uc_translation
}
\description{
A translation table for the Ultimate Consequences data project,
which documents deaths in Bolivian political conflict.
}
\details{
uc_translation <- \code{list(protest_domain = domain_trans_table, pres_admin = pres_trans_table, dec_affiliation = affil_trans_table, perp_affiliation = affil_trans_table, cause_death = cause_trans_table, state_responsibility = sresp_trans_table, intentionality = intent_trans_table, event_title = event_trans_table)}

Note that the duplicate between two variables is intentional,
since these two variables (for the affiliation of the perpetrator
and victim of political violence) use the same set of possible categories.

Source: \url{https://ultimateconsequences.github.io/}
}
\keyword{datasets}
