% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uc_var_table}
\alias{uc_var_table}
\title{Ultimate Consequences Translation Table}
\format{
A tibble with three rows and 63 columns. Each row corresponds
to a single language, "en" (English), "es" (Spanish), and "r_variable"
(the variable names in R).
}
\source{
<https://ultimateconsequences.github.io/>
}
\usage{
uc_var_table
}
\description{
A translation table for the Ultimate Consequences data project,
which documents deaths in Bolivian political conflict.
}
\keyword{datasets}
