% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_name.R
\name{variable_names_vector}
\alias{variable_names_vector}
\title{Produce a list of display names for each variables}
\usage{
variable_names_vector(dataframe)
}
\arguments{
\item{dataframe}{Dataframe whose columns are the variables to be listed.}
}
\value{
A vector of strings wit the display names.
}
\description{
Using the active variable name table, produce a list of display names for each
variable, in the language currently specified as tcats$title_lang (set by
`set_title_lang()`).
}
\examples{
variable_names_vector(forcats::gss_cat)
}
